+++ src/engine/mod.rs
@@ -1,10 +1,7 @@
 //! Provides the [Engine] abstraction and out of the box implementations.
 #[cfg(any(feature = "alloc", test))]
 use crate::chunked_encoder;
-use crate::{
-    encode::{encode_with_padding, EncodeSliceError},
-    encoded_len, DecodeError, DecodeSliceError,
-};
+use crate::{alphabet, encode::{encode_with_padding, EncodeSliceError}, encoded_len, DecodeError, DecodeSliceError};
 #[cfg(any(feature = "alloc", test))]
 use alloc::vec::Vec;
 
@@ -29,6 +26,7 @@ mod naive;
 mod tests;
 
 pub use general_purpose::{GeneralPurpose, GeneralPurposeConfig, Scalar};
+pub use alphabet::Symbol;
 
 /// The runtime-detected SIMD engine. Requires the `simd-unsafe` feature.
 #[cfg(all(
@@ -113,7 +111,7 @@ pub trait Engine: Send + Sync {
     ///
     /// Decoding must not write any bytes into the output slice other than the decoded data.
     ///
-    /// Non-canonical trailing bits in the final tokens or non-canonical padding must be reported as
+    /// Non-canonical trailing bits in the final symbols or non-canonical padding must be reported as
     /// errors unless the engine is configured otherwise.
     #[doc(hidden)]
     fn internal_decode(
@@ -448,6 +446,11 @@ pub trait Engine: Send + Sync {
 
         inner(self, input.as_ref(), output)
     }
+
+    /// Returns the symbol used for encode padding.
+    ///
+    /// Typically this is `'='`, but weird alphabets may use other values.
+    fn padding(&self) -> Symbol;
 }
 
 /// The minimal level of configuration that engines must support.
@@ -473,7 +476,7 @@ pub trait DecodeEstimate {
     /// for pre-allocating buffers, etc.
     ///
     /// The estimate must be no larger than the next largest complete triple of decoded bytes.
-    /// That is, the final quad of tokens to decode may be assumed to be complete with no padding.
+    /// That is, the final quad of symbols to decode may be assumed to be complete with no padding.
     fn decoded_len_estimate(&self) -> usize;
 }